refactor: SecretKey 변수 이름 변경 및 signWith() 함수 파라미터 변경 - #100
