+++ src/main/java/com/example/mutsideout_mju/authentication/token/AccessTokenProvider.java
@@ -1,4 +1,4 @@
-package com.example.mutsideout_mju.authentication;
+package com.example.mutsideout_mju.authentication.token;
 
 import com.example.mutsideout_mju.exception.UnauthorizedException;
 import com.example.mutsideout_mju.exception.errorCode.ErrorCode;
@@ -11,14 +11,15 @@ import org.springframework.stereotype.Component;
 import javax.crypto.SecretKey;
 import java.nio.charset.StandardCharsets;
 import java.util.Date;
+
 @Slf4j
 @Component
 public class AccessTokenProvider {
-    private final SecretKey key; // 시크릿 키
-    private final long validityInMilliseconds; // 유효 시간
+    private final SecretKey key; // AccessToken 시크릿 키
+    private final long validityInMilliseconds; // AccessToken 유효 시간
 
-    public AccessTokenProvider(@Value("${security.jwt.token.secret-key}") final String secretKey,
-                               @Value("${security.jwt.token.expire-length}") final long validityInMilliseconds) {
+    public AccessTokenProvider(@Value("${security.jwt.token.secret-access-key}") final String secretKey,
+                               @Value("${security.jwt.token.access-expire-length}") final long validityInMilliseconds) {
         this.key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
         this.validityInMilliseconds = validityInMilliseconds;
     }
@@ -32,20 +33,19 @@ public class AccessTokenProvider {
                 .setSubject(payload)
                 .setIssuedAt(now)
                 .setExpiration(expiration)
-                .signWith(SignatureAlgorithm.HS256, key)
+                .signWith(key, SignatureAlgorithm.HS256)
                 .compact();
     }
 
     // 정보 추출
     public String getPayload(final String token) {
         try {
-            String payload = Jwts.parserBuilder()
+            return Jwts.parserBuilder()
                     .setSigningKey(key)
                     .build()
                     .parseClaimsJws(token)
                     .getBody()
                     .getSubject();
-            return payload;
         } catch (JwtException e) {
             throw new UnauthorizedException(ErrorCode.INVALID_TOKEN, e.getMessage());
         }
