+++ src/main/java/com/example/mutsideout_mju/authentication/token/RefreshTokenProvider.java
@@ -1,4 +1,4 @@
-package com.example.mutsideout_mju.authentication;
+package com.example.mutsideout_mju.authentication.token;
 
 import javax.crypto.SecretKey;
 
@@ -8,18 +8,17 @@ import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+
 import java.nio.charset.StandardCharsets;
-import java.time.Duration;
 import java.util.Date;
 
 @Component
 public class RefreshTokenProvider {
-    // JwtTokenProvider 키와 다른 키 사용.
-    private final SecretKey key; // 시크릿 키
-    private final long validityInMilliseconds; // 유효 시간
+    private final SecretKey key; // RefreshToken 시크릿 키
+    private final long validityInMilliseconds; // RefreshToken 유효 시간
 
     public RefreshTokenProvider(@Value("${security.jwt.token.secret-refresh-key}") final String secretKey,
-                            @Value("${security.jwt.token.expire-length}") final long validityInMilliseconds) {
+                                @Value("${security.jwt.token.refresh-expire-length}") final long validityInMilliseconds) {
         this.key = Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8));
         this.validityInMilliseconds = validityInMilliseconds;
     }
@@ -27,12 +26,12 @@ public class RefreshTokenProvider {
     // RefreshToken 생성
     public String createRefreshToken() {
         Date now = new Date();
-        Date validity = new Date(now.getTime() + Duration.ofDays(14).toMillis());
+        Date validity = new Date(now.getTime() + validityInMilliseconds);
 
         return Jwts.builder()
                 .setIssuedAt(now)
                 .setExpiration(validity)
-                .signWith(SignatureAlgorithm.HS256, key)
+                .signWith(key, SignatureAlgorithm.HS256)
                 .compact();
     }
